requirements.txt
langchain
python-dotenv
langchain-openai
beautifulsoup4
langchain_community
faiss-cpu
streamlit